+++ src/pages/Transfer.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './Pages.css';
-import Order from 'components/Order';
+import Order from 'components/order/Order';
 
 const Transfer = () =>
   <section className="form-order">
